requirements.txt
# Cow Teat Image Classification - Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
Pillow>=9.5.0
scipy>=1.10.0
scikit-learn>=1.2.0
